requirements/base.txt
# File: requirements/base.txt
boto3==1.34.38
django-ratelimit>=4.0
django==5.0.6
django-environ==0.11.2
django-storages==1.13.2
psycopg2-binary==2.9.9
stream-chat>=3.0
whitenoise==6.6.0
